Split a test file into two.
